muxdb/internal/trie: changes due to trie pkg updates
* add Stage method to replace Hash & Commit
* remove lazy-load code
* no return error on Copy
* add targetCommitNum param for DumpLeaves method

+++ muxdb/muxdb.go
@@ -44,7 +44,7 @@ type Options struct {
 	// TrieRootCacheCapacity is the capacity of the cache for trie root nodes.
 	TrieRootCacheCapacity int
 	// TrieCachedNodeTTL defines the life time(times of commit) of cached trie nodes.
-	TrieCachedNodeTTL int
+	TrieCachedNodeTTL uint16
 	// TrieLeafBankSlotCapacity defines max count of cached slot for leaf bank.
 	TrieLeafBankSlotCapacity int
 	// TrieHistPartitionFactor is the partition factor for historical trie nodes.
